extract client and refactor access token
